Improve test progress tracker
- Use doubles instead of floats, so even big ints can be represented and
compared correctly when checking if we should record progress
- Report stall exceeded only if stall exceeded AND no progress was made
- Introduce configurable stall tolerance with default 20s

+++ hazelcast/src/test/java/com/hazelcast/test/DefaultTaskProgress.java
@@ -20,13 +20,13 @@ public class DefaultTaskProgress implements TaskProgress {
     private final long timestamp = System.currentTimeMillis();
     private final int total;
     private final int done;
-    private final float progress;
+    private final double progress;
     private final boolean completed;
 
     public DefaultTaskProgress(int total, int done) {
         this.total = total;
         this.done = done;
-        this.progress = ((float) done) / total;
+        this.progress = ((double) done) / total;
         this.completed = total == done;
     }
 
@@ -36,7 +36,7 @@ public class DefaultTaskProgress implements TaskProgress {
     }
 
     @Override
-    public float progress() {
+    public double progress() {
         return progress;
     }
